refactor: move sellItem to its own file
