+++ src/game-logic/reducers/index.test.js
@@ -1,4 +1,4 @@
-import { testCrop, testItem } from '../../test-utils'
+import { testCrop } from '../../test-utils'
 import {
   ACHIEVEMENT_COMPLETED,
   LOAN_INCREASED,
@@ -93,142 +93,6 @@ describe('makeRecipe', () => {
         { id: 'sample-item-1', quantity: 1 },
         { id: 'sample-recipe-1', quantity: 1 },
       ])
-    })
-  })
-})
-
-describe('sellItem', () => {
-  test('sells item', () => {
-    const state = fn.sellItem(
-      {
-        inventory: [testItem({ id: 'sample-item-1', quantity: 1 })],
-        itemsSold: {},
-        loanBalance: 0,
-        money: 100,
-        pendingPeerMessages: [],
-        todaysNotifications: [],
-        revenue: 0,
-        todaysRevenue: 0,
-        valueAdjustments: { 'sample-item-1': 1 },
-      },
-      testItem({ id: 'sample-item-1' })
-    )
-
-    expect(state.inventory).toEqual([])
-    expect(state.money).toEqual(101)
-    expect(state.revenue).toEqual(1)
-    expect(state.todaysRevenue).toEqual(1)
-    expect(state.itemsSold).toEqual({ 'sample-item-1': 1 })
-  })
-
-  test('updates learnedRecipes', () => {
-    const { learnedRecipes } = fn.sellItem(
-      {
-        inventory: [testItem({ id: 'sample-item-1', quantity: 3 })],
-        itemsSold: {},
-        loanBalance: 0,
-        money: 100,
-        pendingPeerMessages: [],
-        todaysNotifications: [],
-        revenue: 0,
-        todaysRevenue: 0,
-        valueAdjustments: { 'sample-item-1': 1 },
-      },
-      testItem({ id: 'sample-item-1' }),
-      3
-    )
-
-    expect(learnedRecipes['sample-recipe-1']).toBeTruthy()
-  })
-
-  describe('there is an outstanding loan', () => {
-    let state
-
-    describe('item is not a farm product', () => {
-      test('sale is not garnished', () => {
-        state = fn.sellItem(
-          {
-            inventory: [testItem({ id: 'sample-crop-seeds-1', quantity: 3 })],
-            itemsSold: {},
-            loanBalance: 100,
-            money: 100,
-            pendingPeerMessages: [],
-            todaysNotifications: [],
-            revenue: 0,
-            todaysRevenue: 0,
-            valueAdjustments: { 'sample-crop-seeds-1': 10 },
-          },
-          testItem({ id: 'sample-crop-seeds-1' }),
-          3
-        )
-
-        expect(state.loanBalance).toEqual(100)
-        expect(state.money).toEqual(130)
-        expect(state.revenue).toEqual(30)
-        expect(state.todaysRevenue).toEqual(30)
-      })
-    })
-
-    describe('item is a farm product', () => {
-      describe('loan is greater than garnishment', () => {
-        test('sale is garnished', () => {
-          state = fn.sellItem(
-            {
-              inventory: [testItem({ id: 'sample-crop-1', quantity: 3 })],
-              itemsSold: {},
-              loanBalance: 100,
-              money: 100,
-              pendingPeerMessages: [],
-              todaysNotifications: [],
-              revenue: 0,
-              todaysRevenue: 0,
-              valueAdjustments: { 'sample-crop-1': 10 },
-            },
-            testItem({ id: 'sample-crop-1' }),
-            3
-          )
-
-          expect(state.loanBalance).toEqual(97)
-          expect(state.money).toEqual(157)
-          expect(state.revenue).toEqual(57)
-          expect(state.todaysRevenue).toEqual(57)
-        })
-      })
-
-      describe('loan is less than garnishment', () => {
-        beforeEach(() => {
-          state = fn.sellItem(
-            {
-              inventory: [testItem({ id: 'sample-crop-1', quantity: 3 })],
-              itemsSold: {},
-              loanBalance: 1.5,
-              money: 100,
-              pendingPeerMessages: [],
-              todaysNotifications: [],
-              revenue: 0,
-              todaysRevenue: 0,
-              valueAdjustments: { 'sample-crop-1': 10 },
-            },
-            testItem({ id: 'sample-crop-1' }),
-            3
-          )
-        })
-
-        test('loan is payed off', () => {
-          expect(state.loanBalance).toEqual(0)
-        })
-
-        test('sale is reduced based on remaining loan balance', () => {
-          expect(state.money).toEqual(158.5)
-          expect(state.todaysRevenue).toEqual(58.5)
-        })
-
-        test('payoff notification is shown', () => {
-          expect(state.todaysNotifications).toEqual([
-            { message: LOAN_PAYOFF``, severity: 'success' },
-          ])
-        })
-      })
     })
   })
 })
